magento/magento2#29372: Wrong tags are set for CategoryList query
Split categoryList cache test classes
